add initial support for AWS SDK v2 (classpath still requires fixing)
