+++ pom.xml
@@ -88,6 +88,12 @@
             <artifactId>aws-java-sdk</artifactId>
             <version>${aws.sdk.version}</version>
             <scope>provided</scope>
+            <exclusions>
+              <exclusion>
+                <groupId>com.amazonaws</groupId>
+                <artifactId>aws-java-sdk-kinesisvideo</artifactId>
+              </exclusion>
+            </exclusions>
         </dependency>
         <dependency>
             <groupId>com.amazonaws</groupId>
@@ -113,6 +119,23 @@
             <groupId>com.amazonaws</groupId>
             <artifactId>aws-java-sdk-sqs</artifactId>
             <version>${aws.sdk.version}</version>
+        </dependency>
+
+        <!-- AWS SDK version 2 libs -->
+        <dependency>
+            <groupId>software.amazon.awssdk</groupId>
+            <artifactId>sqs</artifactId>
+            <version>2.13.36</version>
+        </dependency>
+        <dependency>
+            <groupId>software.amazon.awssdk</groupId>
+            <artifactId>kinesis</artifactId>
+            <version>2.13.36</version>
+        </dependency>
+        <dependency>
+            <artifactId>netty-nio-client</artifactId>
+            <groupId>software.amazon.awssdk</groupId>
+            <version>2.0.0</version>
         </dependency>
