+++ src/main/java/cloud/localstack/awssdkv1/TestUtils.java
@@ -1,4 +1,4 @@
-package cloud.localstack;
+package cloud.localstack.awssdkv1;
 
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSCredentialsProvider;
@@ -33,36 +33,16 @@ import com.amazonaws.services.secretsmanager.AWSSecretsManager;
 import com.amazonaws.services.secretsmanager.AWSSecretsManagerClientBuilder;
 import com.amazonaws.services.sqs.AmazonSQS;
 import com.amazonaws.services.sqs.AmazonSQSClientBuilder;
-import java.io.IOException;
-import java.lang.reflect.Field;
-import java.nio.channels.FileChannel;
-import java.nio.file.CopyOption;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardCopyOption;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.stream.Stream;
+
+import cloud.localstack.Constants;
+import cloud.localstack.Localstack;
+import cloud.localstack.CommonUtils;
 
 @SuppressWarnings("all")
 public class TestUtils {
 
-    public static final String DEFAULT_REGION = "us-east-1";
-    public static final String TEST_ACCESS_KEY = "test";
-    public static final String TEST_SECRET_KEY = "test";
-    public static final AWSCredentials TEST_CREDENTIALS = new BasicAWSCredentials(TEST_ACCESS_KEY, TEST_SECRET_KEY);
-
-    private static final String[] EXCLUDED_DIRECTORIES = {
-      ".github", ".git", ".idea", ".venv", "target", "node_modules"
-    };
-
-    public static void setEnv(String key, String value) {
-        Map<String, String> newEnv = new HashMap<String, String>(System.getenv());
-        newEnv.put(key, value);
-        setEnv(newEnv);
-    }
+    public static final AWSCredentials TEST_CREDENTIALS = new BasicAWSCredentials(
+        Constants.TEST_ACCESS_KEY, Constants.TEST_SECRET_KEY);
 
     public static AmazonSQS getClientSQS() {
         return getClientSQS(null);
@@ -227,87 +207,16 @@ public class TestUtils {
         return getEndpointConfiguration(Localstack.INSTANCE.getEndpointStepFunctions());
     }
 
-    protected static void setEnv(Map<String, String> newEnv) {
-        try {
-            Class<?> processEnvironmentClass = Class.forName("java.lang.ProcessEnvironment");
-            Field theEnvironmentField = processEnvironmentClass.getDeclaredField("theEnvironment");
-            theEnvironmentField.setAccessible(true);
-            Map<String, String> env = (Map<String, String>) theEnvironmentField.get(null);
-            env.putAll(newEnv);
-            Field theCaseInsensitiveEnvironmentField = processEnvironmentClass
-                    .getDeclaredField("theCaseInsensitiveEnvironment");
-            theCaseInsensitiveEnvironmentField.setAccessible(true);
-            Map<String, String> cienv = (Map<String, String>) theCaseInsensitiveEnvironmentField.get(null);
-            cienv.putAll(newEnv);
-        } catch (NoSuchFieldException e) {
-            try {
-                Class[] classes = Collections.class.getDeclaredClasses();
-                Map<String, String> env = System.getenv();
-                for (Class cl : classes) {
-                    if ("java.util.Collections$UnmodifiableMap".equals(cl.getName())) {
-                        Field field = cl.getDeclaredField("m");
-                        field.setAccessible(true);
-                        Object obj = field.get(env);
-                        Map<String, String> map = (Map<String, String>) obj;
-                        map.clear();
-                        map.putAll(newEnv);
-                    }
-                }
-            } catch (Exception e2) {
-                e2.printStackTrace();
-            }
-        } catch (Exception e1) {
-            e1.printStackTrace();
-        }
-    }
-
-    public static void disableSslCertChecking() {
-        System.setProperty("com.amazonaws.sdk.disableCertChecking", "true");
-    }
-
-    public static void copyFolder(Path src, Path dest) throws IOException {
-        try(Stream<Path> stream = Files.walk(src)) {
-            stream.forEach(source -> {
-                boolean isExcluded = Arrays.stream(EXCLUDED_DIRECTORIES)
-                    .anyMatch( excluded -> source.toAbsolutePath().toString().contains(excluded));
-                if (!isExcluded) {
-                    copy(source, dest.resolve(src.relativize(source)));
-                }
-            });
-        }
-    }
-
-    public static void copy(Path source, Path dest) {
-        try {
-            CopyOption[] options = new CopyOption[] {StandardCopyOption.COPY_ATTRIBUTES, StandardCopyOption.REPLACE_EXISTING};
-            if(Files.isDirectory(dest)) {
-                // continue without copying
-                return;
-            }
-            if (Files.exists(dest)) {
-                try(FileChannel sourceFile = FileChannel.open(source)) {
-                    try (FileChannel destFile = FileChannel.open(dest)) {
-                        if (!Files.getLastModifiedTime(source).equals(Files.getLastModifiedTime(dest))
-                                || sourceFile.size() != destFile.size()
-                        ) {
-                            Files.copy(source, dest, options);
-                        }
-                    }
-                }
-            } else {
-                Files.copy(source, dest, options);
-            }
-        } catch (Exception e) {
-            throw new RuntimeException(e.getMessage(), e);
-        }
-    }
+    /**
+     * UTIL METHODS
+     */
 
     public static AWSCredentialsProvider getCredentialsProvider() {
         return new AWSStaticCredentialsProvider(TEST_CREDENTIALS);
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfiguration(String endpointURL) {
-        return new AwsClientBuilder.EndpointConfiguration(endpointURL, DEFAULT_REGION);
+        return new AwsClientBuilder.EndpointConfiguration(endpointURL, Constants.DEFAULT_REGION);
     }
 
 }
