+++ src/main/java/cloud/localstack/deprecated/TestUtils.java
@@ -38,7 +38,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
 
-import static cloud.localstack.TestUtils.getCredentialsProvider;
+import static cloud.localstack.awssdkv1.TestUtils.getCredentialsProvider;
 
 @Deprecated
 @SuppressWarnings("all")
